feat(ui)!: replace Win32/GTK windowing with unified SDL3 backend
