+++ cmake/rexglue_helpers.cmake
@@ -39,7 +39,7 @@ endfunction()
 # rexglue_configure_target(<target>) - Host application
 #
 # Adds:
-#   - Platform entry point source (windowed_app_main_*.cpp)
+#   - SDL3 entry point source (windowed_app_main_sdl.cpp)
 #   - ReXApp base class source (rex_app.cpp)
 #   - Build-config define for the version stamp
 #   - $ORIGIN RPATH on UNIX so the host finds librexruntime.so next to itself
@@ -48,15 +48,8 @@ endfunction()
 #     so this single copy handles them transitively.
 #==========================================================
 function(rexglue_configure_target target_name)
-    if(WIN32)
-        target_sources(${target_name} PRIVATE
-            ${REXGLUE_SHARE_DIR}/windowed_app_main_win.cpp)
-    else()
-        target_sources(${target_name} PRIVATE
-            ${REXGLUE_SHARE_DIR}/windowed_app_main_posix.cpp)
-    endif()
-
     target_sources(${target_name} PRIVATE
+        ${REXGLUE_SHARE_DIR}/windowed_app_main_sdl.cpp
         ${REXGLUE_SHARE_DIR}/rex_app.cpp)
 
     target_compile_definitions(${target_name} PRIVATE
