--- src/ui/windowed_app_main_posix.cpp
@@ -1,81 +1,0 @@
-/**
- ******************************************************************************
- * Xenia : Xbox 360 Emulator Research Project                                 *
- ******************************************************************************
- * Copyright 2021 Ben Vanik. All rights reserved.                             *
- * Released under the BSD license - see LICENSE in the root for more details. *
- ******************************************************************************
- *
- * @modified    Tom Clay, 2026 - Adapted for ReXGlue runtime
- */
-
-#include <cstdio>
-#include <cstdlib>
-#include <map>
-#include <memory>
-#include <string>
-#include <vector>
-
-#include <rex/cvar.h>
-#include <rex/logging.h>
-#include <rex/platform/env.h>
-#include <rex/ui/windowed_app.h>
-#include <rex/ui/windowed_app_context_gtk.h>
-
-#include <gtk/gtk.h>
-
-extern "C" int main(int argc_pre_gtk, char** argv_pre_gtk) {
-  // Before touching anything GTK+, make sure that when running on Wayland,
-  // we'll still get an X11 (Xwayland) window
-  rex::platform::env::set("GDK_BACKEND", "x11");
-
-  // Initialize GTK+, which will handle and remove its own arguments from argv.
-  // Both GTK+ and Xenia use --option=value argument format (see man
-  // gtk-options), however, it's meaningless to try to parse the same argument
-  // both as a GTK+ one and as a cvar. Make GTK+ options take precedence in case
-  // of a name collision, as there's an alternative way of setting Xenia options
-  // (the config).
-  int argc_post_gtk = argc_pre_gtk;
-  char** argv_post_gtk = argv_pre_gtk;
-  gtk_disable_setlocale();
-  if (!gtk_init_check(&argc_post_gtk, &argv_post_gtk)) {
-    // Logging has not been initialized yet.
-    std::fputs("Failed to initialize GTK+\n", stderr);
-    return EXIT_FAILURE;
-  }
-
-  auto remaining = rex::cvar::Init(argc_post_gtk, argv_post_gtk);
-  rex::cvar::ApplyEnvironment();
-  rex::InitLoggingEarly();
-
-  int result;
-
-  {
-    rex::ui::GTKWindowedAppContext app_context;
-
-    std::unique_ptr<rex::ui::WindowedApp> app = rex::ui::GetWindowedAppCreator()(app_context);
-
-    // Match remaining positional args to app's expected options
-    const auto& option_names = app->GetPositionalOptions();
-    std::map<std::string, std::string> parsed;
-    size_t count = std::min(remaining.size(), option_names.size());
-    for (size_t i = 0; i < count; ++i) {
-      parsed[option_names[i]] = remaining[i];
-    }
-    app->SetParsedArguments(std::move(parsed));
-
-    if (app->OnInitialize()) {
-      app_context.RunMainGTKLoop();
-      result = EXIT_SUCCESS;
-    } else {
-      result = EXIT_FAILURE;
-    }
-
-    app->InvokeOnDestroy();
-  }
-
-  // Logging may still be needed in the destructors.
-  rex::ShutdownLogging();
-
-  return result;
-}
